+++ pkgs/native_assets_builder/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.16.0-wip
+
+- Pass in path to pubspec to read user-defines (and read user-defines) in this
+  package rather than in the SDKs using this package.
+
 ## 0.15.0
 
 - Bump `package:native_assets_cli` to 0.15.0.